fix: Claude-led code audit — 13 fixes across 8 files (3 critical, 6 high, 4 medium)
CRITICAL:
- slop_mesh.h: fix buffer overflow in sendTextTo (strlen after strncpy truncation)
- mesh_wrapper.cpp: add null terminators after strncpy (UB on exact-fit strings)
- keyboard.cpp: change Wire.read() result from char to int (0xFF indistinguishable from -1)

HIGH:
- mesh_wrapper.cpp: add null check after new SlopMesh()
- keyboard.cpp: check Wire.endTransmission() return values in init
- chat_screen.cpp: make back button and channel prev/next clickable (were plain labels)
- display.cpp: add lv_indev_set_group + default group for keyboard input routing
- ui.cpp: enable live battery/signal status bar updates + periodic saveState()
- tdeck_board.h: power down peripherals before deep sleep (battery drain fix)

MEDIUM:
- touch.cpp: static point_data buffer (avoids repeated 40-byte stack alloc)
- keyboard.cpp: implement modifier tracking (shift via case, alt via 0x0C toggle)
- sdcard.cpp: fix %llu printf format (cast to unsigned long for ESP32 compat)
- display.cpp: named constant LVGL_DRAW_BUF_LINES replacing magic number 20

Tests: 160/161 pass | ESP32 build: 998KB flash, 114KB RAM | No tag

+++ src/hal/sdcard.cpp
@@ -76,10 +76,10 @@ const char* slopos_sdcard_format_size(uint64_t bytes, char* buf, size_t buf_sz)
         snprintf(buf, buf_sz, "%.1f MB",
                  (double)bytes / (1024.0 * 1024.0));
     } else if (bytes >= 1024) {
-        snprintf(buf, buf_sz, "%llu KB",
-                 (unsigned long long)(bytes / 1024));
+        snprintf(buf, buf_sz, "%lu KB",
+                 (unsigned long)(bytes / 1024));
     } else {
-        snprintf(buf, buf_sz, "%llu B", (unsigned long long)bytes);
+        snprintf(buf, buf_sz, "%lu B", (unsigned long)bytes);
     }
     return buf;
 }
